Adds product creation form
Implements a new product creation page and form using React Hook Form and Zod for validation.

This introduces a dedicated component for managing product creation with fields for title, subtitle, handle, and description. It also includes auto-generation of the handle from the title.

The form integrates with a rich text editor for product descriptions and uses toast notifications to display success or error messages upon submission. It includes client-side validation and redirects to the product listing page after successful creation.

Also restyles the footer and navigation for improved user experience.